+++ lib/l10n/app_localizations_de.dart
@@ -5482,6 +5482,9 @@ class AppLocalizationsDe extends AppLocalizations {
   String get journalLinkedEntriesActivityFilterAudio => 'Audio';
 
   @override
+  String get journalLinkedEntriesActivityFilterCode => 'Code';
+
+  @override
   String get journalLinkedEntriesActivityFilterImages => 'Bilder';
 
   @override